Created a different component for promotions list and removed <Toolbar/>, <Footer/>
